python_script/requirements.txt
beautifulsoup4
pandas
python-dotenv
requests
selenium
supabase
unidecode
